backend/requirements.txt
fastapi
uvicorn
pymongo
boto3
scapy
python-dotenv
motor
certifi
websockets
psutil
